add user, move mock-data file and add loading
