+++ src/app/pages/employees/employee-list/employee-list.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { Table, TableModule } from 'primeng/table';
-import { employeesData } from '../mock-data/mock-employees';
+import { employeesData } from '../../mock-data/mock-employees';
 import { ButtonModule } from 'primeng/button';
 import { MessageService } from 'primeng/api';
 import { ToastModule } from 'primeng/toast';
