+++ src/app/pages/mock-data/mock-users.ts
@@ -1,0 +1,17 @@
+export const users = [
+  {
+    "username": "admin",
+    "password": "1234",
+    "role": "admin"
+  },
+  {
+    "username": "emp1",
+    "password": "4567",
+    "role": "employee"
+  },
+  {
+    "username": "emp2",
+    "password": "7890",
+    "role": "employee"
+  }
+]
